Extract BranchSelector into separate component for composability
Moves branch selection state, data fetching, filtering logic, and UI
(~210 lines) out of CreateTaskModal into a reusable BranchSelector component.
CreateTaskModal drops from 507 to 295 lines.
